fix(design-system): move files to remove circular module references
