+++ packages/web-components/fast-foundation/src/design-system/index.ts
@@ -1,2 +1,3 @@
 export * from "./design-system";
 export * from "./component-presentation";
+export * from "./registration-context";
